Move TimeValuePair to Utility, Ring no longer inherits from VariableSizeModule
Also migrated the old value for curve to scaleCurve.
