+++ NewHorizons/Builder/Body/ProxyBuilder.cs
@@ -188,7 +188,7 @@ namespace NewHorizons.Builder.Body
             newProxy.SetActive(true);
         }
 
-        private static GameObject AddColouredSphere(GameObject rootObj, float size, VariableSizeModule.TimeValuePair[] curve, Color color)
+        private static GameObject AddColouredSphere(GameObject rootObj, float size, TimeValuePair[] curve, Color color)
         {
             GameObject sphereGO = GameObject.CreatePrimitive(PrimitiveType.Sphere);
             sphereGO.transform.name = "ProxySphere";
@@ -206,7 +206,7 @@ namespace NewHorizons.Builder.Body
             return sphereGO;
         }
 
-        private static void AddSizeController(GameObject go, VariableSizeModule.TimeValuePair[] curve, float size)
+        private static void AddSizeController(GameObject go, TimeValuePair[] curve, float size)
         {
             var sizeController = go.AddComponent<SizeController>();
             sizeController.SetScaleCurve(curve);
