+++ NewHorizons/Components/SizeControllers/SizeController.cs
@@ -1,4 +1,5 @@
 using NewHorizons.External.Modules.VariableSize;
+using NewHorizons.Utility;
 using UnityEngine;
 namespace NewHorizons.Components.SizeControllers
 {
@@ -22,7 +23,7 @@ namespace NewHorizons.Components.SizeControllers
             base.transform.localScale = Vector3.one * CurrentScale;
         }
 
-        public void SetScaleCurve(VariableSizeModule.TimeValuePair[] curve)
+        public void SetScaleCurve(TimeValuePair[] curve)
         {
             scaleCurve = new AnimationCurve();
             foreach (var pair in curve)
