+++ NewHorizons/External/Modules/VariableSize/VariableSizeModule.cs
@@ -1,3 +1,4 @@
+using NewHorizons.Utility;
 using Newtonsoft.Json;
 using UnityEngine;
 
@@ -7,22 +8,8 @@ namespace NewHorizons.External.Modules.VariableSize
     public class VariableSizeModule
     {
         /// <summary>
-        /// Scale this module over time
+        /// Scale this object over time
         /// </summary>
         public TimeValuePair[] curve;
-
-        [JsonObject]
-        public class TimeValuePair
-        {
-            /// <summary>
-            /// A specific point in time
-            /// </summary>
-            public float time;
-
-            /// <summary>
-            /// The value for this point in time
-            /// </summary>
-            public float value;
-        }
     }
 }
